Move compilation threadpool from MeshDevice to MetalContext
