+++ ttnn/cpp/ttnn-pybind/device.cpp
@@ -78,19 +78,12 @@ void ttnn_device(py::module& module) {
 
     module.def(
         "wait_for_compile_threadpool",
-        [](ttnn::MeshDevice& device) { device.wait_for_kernel_compilation_thread_pool(); },
-        py::arg("device"),
+        []() { tt::tt_metal::detail::WaitForKernelCompilationThreadPool(); },
         R"doc(
-        Wait for all operations in the device thread pool to complete.
-
-        Args:
-            device (ttnn.Device): The device to wait for.
+        Wait for all operations in the kernel compilation thread pool to complete.
 
         Example:
-            >>> device_id = 0
-            >>> device = ttnn.open_device(device_id=device_id)
-            >>> # Queue some operations on the device
-            >>> ttnn.wait_for_threadpool(device)
+            >>> ttnn.wait_for_compile_threadpool()
     )doc");
 }
